Close Grid row click simulation - comment closing div as common issue
Letting indexes allow you to click on the row content and then view
the record. The affected rows are nested within a div with a
data-link attribute. The nesting should end before the other row
links as they are pointing to a different end point. Having two
possible end points for one click leads to a bad user experience!